refactor: move Luarock class into separate file
