+++ rocks-lib/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod build;
 pub mod config;
+pub mod luarock;
 pub mod manifest;
 pub mod rocks;
 pub mod rockspec;
